Refactor utilities and folder structure
